% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curefit.R
\name{curefit}
\alias{curefit}
\title{Estimated cure proportion}
\usage{
curefit(x, covarprofile)
}
\arguments{
\item{x}{A lpsmc object.}

\item{covarprofile}{The covariate profile on which to compute the
cure proportion.}
}
\value{
A table with the estimated cure proportion.
}
\description{
Computes the estimated cure proportion based on a mixture cure model fit
with \code{lpsmc}. Point estimates and approximate 90% and 95% credible
intervals are shown.
}
\examples{
### Application on breast cancer data
rm(list=ls())
data("breastcancer")
formula <- Surv(tobs, delta) ~ inci(AGE + ER) + late(AGE + ER)
fitcancer <- lpsmc(formula = formula, data = breastcancer, K = 20)
covarprofile <- matrix(c(1, 30, 1, 1, 40, 0), nrow = 2 , byrow = TRUE)
fitcure <- curefit(fitcancer, covarprofile)
fitcure$estimcure

}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr} .
}
