% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LangevinGibbs.R
\name{LangevinGibbs}
\alias{LangevinGibbs}
\title{Langevin-Gibbs sampler for inference in a mixture cure model.}
\usage{
LangevinGibbs(
  formula,
  data,
  K = 15,
  penorder = 3,
  deltaprior = 1e-04,
  mcmcsample = 10000,
  burnin = 2000,
  tunparam = 0.25,
  mcmcseed = NULL,
  progbar = c("yes", "no")
)
}
\arguments{
\item{formula}{A model formula of the form \code{Surv(tobs,delta)~
inci()+late()}.}

\item{data}{A data frame.}

\item{K}{The number of B-spline coefficients.}

\item{penorder}{The order of the penalty associated to the B-spline
coefficients.}

\item{deltaprior}{The parameters of the Gamma prior for the dispersion
parameter.}

\item{mcmcsample}{The length of the MCMC chain.}

\item{burnin}{The length of the burnin.}

\item{tunparam}{The initial tuning parameter of the MLWG algorithm,
default is 0.25.}

\item{mcmcseed}{The seed to be used (for reproducibility).}

\item{progbar}{Should a progress bar be shown? Default is yes.}
}
\description{
This routine implements a Metropolis-Langevin-within-Gibbs sampler to draw
samples from the posterior distribution of a mixture cure model. A
Metropolis-adjusted Langevin algorithm is used to sample from the
conditional posterior distribution of the latent vector. MCMC samples for
the roughness penalty parameter and the dispersion parameter are obtained
via a Gibbs sampler.
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr} .
}
