% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpsmc_object.R
\name{lpsmc.object}
\alias{lpsmc.object}
\title{Objects related to lpsmc routine.}
\value{
\item{X}{The matrix of covariates in the incidence part (including
 an intercept). As the model includes an intercept, the first column
 corresponds to a column of ones.}
\item{Z}{The matrix of covariates in the latency part.}
\item{latenthat}{The vector of estimated latent variables. This includes
the vector of B-spline coefficients, the regression coefficients
of the incidence part and latency part respectively.}
\item{thetahat}{The vector of estimated B-spline coefficients.}
\item{betahat}{The vector of estimated coefficients in the incidence part.}
\item{gammahat}{The vector of estimated coefficents in the latency part.}
\item{vhat}{The (approximate) posterior maximum for the penalty parameter
(in log scale).}
\item{tu}{The largest observed follow-up time.}
\item{ftime}{The vector of observed survival times.}
\item{penorder}{The chosen penalty order for P-splines.}
\item{K}{The number of B-spline coefficients.}
\item{p}{The number of regression in the incidence part (including
intercept).}
\item{q}{The number of regression coefficients in the latency part.}
\item{xmean}{The mean covariate vector for the incidence part.}
\item{px}{The logistic link function}
\item{cumult}{A function required to compute the estimated baseline
survival}
\item{CI90}{The (approximate) 90% credible intervals for the regression
coefficients in the incidence and latency part.}
\item{CI95}{The (approximate) 95% credible intervals for the regression
coefficients in the incidence and latency part.}
\item{CIcure90}{The (approximate) 90% credible interval for the
cure rate (at the mean covariate profile).}
\item{CIcure95}{The (approximate) 95% credible interval for the
cure rate (at the mean covariate profile).}
\item{CIincid90}{The (approximate) 90% credible interval for the
incidence rate (at the mean covariate profile).}
\item{CIincid95}{The (approximate) 95% credible interval for the
incidence rate (at the mean covariate profile).}
}
\description{
A list of objects resulting from the \link{lpsmc} routine.
}
\seealso{
\link{lpsmc}.
}
\author{
Oswaldo Gressani \email{oswado_gressani@hotmail.fr} .
}
