% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpsmc.R
\name{lpsmc}
\alias{lpsmc}
\title{Fit a mixture cure model with Laplacian-P-splines}
\usage{
lpsmc(formula, data, K = 15, penorder = 3, stepsize = 0.2)
}
\arguments{
\item{formula}{A model formula of the form \code{Surv(tobs,delta)~
inci()+late()}.}

\item{data}{A data frame (optional).}

\item{K}{The number of B-spline coefficients.}

\item{penorder}{The order of the penalty.}

\item{stepsize}{The stepsize taken to maximize the log posterior penalty.}
}
\value{
An object of class \code{lpsmc}.
}
\description{
This routine fits a mixture cure model with a logistic link function for
the incidence part and a flexible Cox proportional hazards model for the
latency part where the baseline survival is approximated with penalized
 B-splines.
}
\examples{
### Real data application ECOG e1684 clinical trial
data("ecog1684")
formula <- Surv(tobs,delta) ~ inci(SEX + TRT + AGE) + late(SEX + TRT + AGE)
fite1684 <- lpsmc(formula = formula, data = ecog1684)
fite1684

### Application on breast cancer data
rm(list=ls())
data("breastcancer")
formula <- Surv(tobs, delta) ~ inci(AGE + ER) + late(AGE + ER)
fitcancer <- lpsmc(formula = formula, data = breastcancer, K = 20)
fitcancer

}
\seealso{
\link{lpsmc.object}
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr} .
}
