% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postpendist.R
\name{postpendist}
\alias{postpendist}
\title{Plot the normalized approximate posterior of the penalty parameter}
\usage{
postpendist(x, low, up, themetype = c("classic", "gray", "light", "dark"))
}
\arguments{
\item{x}{An object of class \code{lpsmc}}

\item{low}{The lower bound on the x-axis (in log scale).}

\item{up}{The upper bound on the x-axis (in log scale).}

\item{themetype}{The theme, either "classic", "gray", "light" or "dark".}
}
\description{
Plots the normalized approximate postrior of the penalty parameter based
on a an object of class \code{lpsmc}.
}
\examples{
### Posterior penalty distribution for breast cancer dataset
rm(list=ls())
data("breastcancer")
formula <- Surv(tobs, delta) ~ inci(AGE + ER) + late(AGE + ER)
fitcancer <- lpsmc(formula = formula, data = breastcancer, K = 20,
                    stepsize = 0.1)
postpendist(fitcancer, 5, 12, themetype = "gray")

}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr} .
}
