% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_LangevinGibbs_print.R
\name{print.LangevinGibbs}
\alias{print.LangevinGibbs}
\title{Print a LangevinGibbs object.}
\usage{
\method{print}{LangevinGibbs}(x, ...)
}
\arguments{
\item{x}{An object of class \code{LangevinGibbs}.}

\item{...}{Further arguments to be passed to print routine.}
}
\description{
Print method for a \code{LangevinGibbs} object.
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr} .
}
