% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_lpsmc_print.R
\name{print.lpsmc}
\alias{print.lpsmc}
\title{Print a lpsmc object.}
\usage{
\method{print}{lpsmc}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lpsmc}.}

\item{...}{Further arguments to be passed to print routine.}
}
\description{
Print method for a \code{lpsmc} object.
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr} .
}
