% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLangevin.R
\name{simLangevin}
\alias{simLangevin}
\title{Assessing the statistical performance of LangevinGibbs.}
\usage{
simLangevin(
  n = 300,
  K = 15,
  scenario = 1,
  S = 500,
  chainlength = 1500,
  burn = 500,
  exactrep = FALSE,
  simnum = NULL
)
}
\arguments{
\item{n}{Sample size.}

\item{K}{Number of B-spline basis functions.}

\item{scenario}{Either 1 or 2.}

\item{S}{Total number of replications.}

\item{chainlength}{The length of the MCMC chain.}

\item{burn}{Burnin length.}

\item{exactrep}{Exactly replicate results.}

\item{simnum}{A seed for reproducibility.}
}
\description{
Simulates the LangevinGibbs routine.
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr} .
}
