% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simlpsmc.R
\name{simlpsmc}
\alias{simlpsmc}
\title{Assessing the statistical performance of lpsmc}
\usage{
simlpsmc(
  n = 300,
  K = 15,
  scenario = 1,
  S = 500,
  exactrep = FALSE,
  simnum = NULL,
  themetype = c("classic", "gray", "light", "dark")
)
}
\arguments{
\item{n}{Sample size.}

\item{K}{Number of B-spline basis functions.}

\item{scenario}{Either 1 or 2.}

\item{S}{Total number of replications.}

\item{exactrep}{Exactly replicate results.}

\item{simnum}{A seed for reproducibility.}

\item{themetype}{The theme of the plot either "classic", "gray","light"
or "dark".}
}
\description{
This routine can be used to assess the statistical performance of the
\code{lpsmc} routine in a simulation setting. The scenario (1 or 2)
determines the data generating process and the total number of replications
can be fixed by the user.
}
\examples{
### Scenario 1,  n=300
sim1 <- simlpsmc(n = 300, scenario = 1, S = 10, exactrep = TRUE)
suppressWarnings(print(sim1$S0plot))
sim1$ASEplot

}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr} .
}
