% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survcurve.R
\name{survcurve}
\alias{survcurve}
\title{Survival curves for the mixture cure model}
\usage{
survcurve(
  x,
  type = c("baseline", "uncured"),
  covarprofile = NULL,
  cred.int = 0.95,
  themetype = c("classic", "gray", "light", "dark"),
  outlisted = FALSE
)
}
\arguments{
\item{x}{An object of type \code{lpsmc}.}

\item{type}{Either baseline or uncured.}

\item{covarprofile}{The matrix of covariate profiles.}

\item{cred.int}{The level for building the credible intervals.}

\item{themetype}{The theme of the plot either "classic", "gray","light"
or "dark".}

\item{outlisted}{Should a list be produced as an output?}
}
\description{
This routine creates objects related to the baseline survival
curve and the survival curve of the uncured subjects in a mixture
cure model fitted with \code{lpsmc}.
}
\examples{
## Plot survival curves of uncured subjects for AGE=(30,50) and ER=1
rm(list=ls())
data("breastcancer")
formula <- Surv(tobs, delta) ~ inci(AGE + ER) + late(AGE + ER)
fitcancer <- lpsmc(formula = formula, data = breastcancer, K = 20)
profilematrix <- matrix(c(30, 1, 50, 1), nrow = 2, byrow = TRUE)
survcurve(fitcancer, type = "uncured", covarprofile = profilematrix)

}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr} .
}
